+++ crates/fuel-core/src/service/adapters/block_importer.rs
@@ -39,7 +39,10 @@ use fuel_core_types::{
         consensus::Consensus,
         SealedBlock,
     },
-    fuel_types::BlockHeight,
+    fuel_types::{
+        BlockHeight,
+        ChainId,
+    },
     services::executor::{
         ExecutionTypes,
         Result as ExecutorResult,
@@ -50,12 +53,13 @@ use std::sync::Arc;
 
 impl BlockImporterAdapter {
     pub fn new(
+        chain_id: ChainId,
         config: Config,
         database: Database,
         executor: ExecutorAdapter,
         verifier: VerifierAdapter,
     ) -> Self {
-        let importer = Importer::new(config, database, executor, verifier);
+        let importer = Importer::new(chain_id, config, database, executor, verifier);
         importer.init_metrics();
         Self {
             block_importer: Arc::new(importer),