+++ crates/services/producer/src/block_producer/gas_price.rs
@@ -1,4 +1,8 @@
-use fuel_core_types::fuel_types::BlockHeight;
+use fuel_core_types::{
+    blockchain::header::ConsensusParametersVersion,
+    fuel_types::BlockHeight,
+};
+use std::sync::Arc;
 
 /// The parameters required to retrieve the gas price for a block
 pub struct GasPriceParams {
@@ -26,4 +30,14 @@ impl From<BlockHeight> for GasPriceParams {
 pub trait GasPriceProvider {
     /// The gas price for all transactions in the block.
     fn gas_price(&self, params: GasPriceParams) -> Option<u64>;
+}
+
+/// Interface for retrieving the consensus parameters.
+#[cfg_attr(feature = "test-helpers", mockall::automock)]
+pub trait ConsensusParametersProvider {
+    /// Retrieve the consensus parameters for the `version`.
+    fn consensus_params_at_version(
+        &self,
+        version: &ConsensusParametersVersion,
+    ) -> Arc<fuel_core_types::fuel_tx::ConsensusParameters>;
 }